Restructure to be more embeddable.
The tiffgl submodule has been removed, requiring an external version or
one higher up in the project tree.

For compiling with an external version outside of your standard paths,
add a Tup config with CCFLAGS and LDFLAGS modified with appropriate -I
and -L.

For compiling in-tree, you must do the above but also set TGL_LIB to the
path to libtgl.a, for order-only dependencies.